[SC-105] refactor: 채팅방 컴포넌트 분리
