+++ src/app/(without-bottom-nav)/chatRoom/[chatRoomId]/page.tsx
@@ -10,7 +10,8 @@ import { Stomp } from '@stomp/stompjs';
 import  SockJS from 'sockjs-client';
 import { deleteRoom } from '@/api/chat/deleteRoom';
 import { PostChatImage } from '@/api/chat/postChatImage';
-import MessageForm from '@/components/MessageForm';
+import MessageForm from '@/components/chat/MessageForm';
+import MessageList from '@/components/chat/MessageList';
 
 // 메시지 타입 정의
 interface Message {
@@ -59,19 +60,7 @@ export default function ChatRoom() {
     const toggleMenu = () => {
         setMenuOpen((prev) => !prev); // 메뉴 열기/닫기
     };
-    const formatCustomDate = (customDate: string) => {
-        try {
-            const [datePart, timePart] = customDate.split(' ');
-            const [hours, minutes] = timePart.split(':').map(Number);
-    
-            const period = hours >= 12 ? '오후' : '오전';
-            const formattedHours = hours % 12 || 12; // 12시간제 변환
-            return `${period} ${formattedHours}:${minutes.toString().padStart(2, '0')}`;
-        } catch (error) {
-            console.error('날짜 변환 오류:', error);
-            return customDate; // 변환 실패 시 원래 값 반환
-        }
-    };
+  
 
     useEffect(() => {
         const token = localStorage.getItem("accessToken");
@@ -175,183 +164,13 @@ const handleScroll = () => {
     }
 };
 
-    const Message = ({ id, content, createdAt, contentType }: Message) => {
-        const messageClass = id != myId ?  'message-left':'message-right' ;
-        return (
-            <div className={messageClass}>
-            {id == myId ? (
-               <div className="modi" />
-           ) : (
-               // me가 아닐 경우에만 profile div를 추가로 표시
-               <div id="profile"></div> // 프로필을 나타내는 div, 필요에 따라 수정 가능
-           )}
-               <div id={id} className={`message_${messageClass}`}>
-                   <div className="message-text">    
-                    {contentType === 'IMAGE' ? (
-                        <img src={content} alt="첨부된 이미지" className="message-image" />
-                    ) : (
-                        content
-                    )}
-                    </div>
-               </div>
-               <div id='time'>{createdAt}</div>
-           </div>
-            
-        );
-    };
-
-    const MessageList = ({ messages }: MessageListProps) => {
-        const renderMessagesWithDateSeparators = () => {
-            const result: JSX.Element[] = [];
-            let lastDate: string | null = null;
-    
-            messages.forEach((message, i) => {
-                const messageDate = message.createdAt.split(' ')[0]; // 메시지 날짜 부분 추출
-                console.log(messageDate);
-                if ( messageDate !== lastDate) {
-                    // 날짜가 바뀌면 새로운 날짜 표시 추가
-                    result.push(
-                        <div id='date' key={`date-${messageDate}-${i}`} className="date-separator">
-                            {messageDate}
-                        </div>
-                    );
-                    lastDate = messageDate;
-                }
-    
-                // 메시지 추가
-                result.push(
-                    <Message
-                        key={i}
-                        id={message.senderId}
-                        content={message.content}
-                        me={message.me}
-                        senderId={''}
-                        createdAt={formatCustomDate(message.createdAt)}
-                        contentType={message.contentType}
-                    />
-                );
-            });
-    
-            return result;
-        };
-    
-        return (
-            <div className="messages" >
-                {renderMessagesWithDateSeparators()}
-                <div ref={messagesEndRef} />
-            </div>
-        );
-    };
+ 
 
     useEffect(() => {
         messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
     }, [messages]);
 
-    // const MessageForm = ({ onMessageSubmit }: MessageFormProps) => {
-    //     const [messageContent, setMessageContent] = useState<string>('');
-    //     const [time, setTime] = useState<string>('');
-    //     const inputRef = useRef<HTMLInputElement | null>(null); 
-    //     const getCurrentTime = (date:Date) => {
-    //         const year = date.getFullYear();
-    //         const month = (date.getMonth() + 1).toString().padStart(2, '0'); // 월은 0부터 시작하므로 1을 더해야 함
-    //         const day = date.getDate().toString().padStart(2, '0');
-    //         const hours = date.getHours().toString().padStart(2, '0');
-    //         const minutes = date.getMinutes().toString().padStart(2, '0');
-    //         const seconds = date.getSeconds().toString().padStart(2, '0');
-        
-    //         return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-    //     };
-
-    //     const handleSubmit = (e: FormEvent) => {
-    //         e.preventDefault();
-
-    //         const currentTime = new Date();
-    //         const formattedTime = getCurrentTime(currentTime)
-            
-    //         const message: Message = {
-    //             content: messageContent,
-    //             me: true,
-    //             senderId: myId,
-    //             id: '',
-    //             createdAt: formattedTime,
-    //             contentType: 'TEXT',
-    //         };
-    //         if (imageFile) {
-    //             // 이미지가 선택되었으면 Base64로 변환하여 메시지에 포함
-    //             const reader = new FileReader();
-    //             reader.onloadend = () => {
-    //               const imageBase64 = reader.result as string;
-    //               const imageMessage: Message = {
-    //                 ...message,
-    //                 content: imageBase64,
-    //                 contentType: 'IMAGE' // 이미지 URL을 추가
-    //               };
-    //               onMessageSubmit(imageMessage); // 이미지 메시지 전송
-    //               PostChatImage(accessToken, imageFile);
-    //               console.log('이미지 파일 보냄');
-    //               setImageFile(null); // 이미지 파일 상태 초기화
-    //             };
-    //             reader.readAsDataURL(imageFile); // 파일을 Base64로 읽음
-    //         } else {   
-    //              onMessageSubmit(message); // 일반 텍스트 메시지 전송
-    //         }
-      
-         
-           
-    //     };
-
-    //     const handleImageUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
-    //         if (e.target.files) {
-    //           const file = e.target.files[0]; 
-    //           setImageFile(file); // 선택한 파일 상태 업데이트
-
-    //           // Base64 변환하여 미리보기 설정
-    //           const reader = new FileReader();
-    //           reader.onloadend = () => {
-    //               setImagePreview(reader.result as string);
-    //           };
-    //           reader.readAsDataURL(file);
-    //         }    
-    //       };
-
-
-    //     return (
-    //      <div id='imagePrevCont'>
-    //                 {imagePreview && (   
-    //         <div className="image-preview">
-    //             <button id='imgPrevDelete' onClick={() => {setImagePreview(null); setImageFile(null)}}>x</button> {/* 미리보기 제거 버튼 */}
-    //          <img id='imagePrev' src={imagePreview} alt="미리보기 이미지" />
-    //          </div>
-    // )}
-    //         <div id='inputCont'>
-            
-    //              <input
-    //       type="file"
-    //       accept="image/*"
-    //       onChange={handleImageUpload}
-    //       style={{ display: 'none' }}
-    //       ref={inputRef}
-    //     />
-    //     <button
-    //       id="imageSubmit"
-    //       onClick={() => inputRef.current?.click()} // 이미지 파일 선택
-    //     ></button>
-    //         <form onSubmit={handleSubmit} id='messagesendForm' autoComplete='off'>
-    //             <input
-    //                 id='messageInput'
-    //                 type="text"
-    //                 value={messageContent}
-    //                 onChange={(e) => setMessageContent(e.target.value)}
-    //                 placeholder="메시지를 입력하세요"
-    //                 autoFocus
-                   
-    //             />
-    //             <button type="submit" id='sendBtn'></button>
-    //         </form>
-    //         </div>
-    //         </div>
-    //     );
-    // };
+   
 
 
     const exitRoom = async(chatRoomId: string | string[] ) => {
@@ -407,7 +226,7 @@ const handleScroll = () => {
            
             <div id='chatBox' ref={chatBoxRef} onScroll={handleScroll}>
             {isLoading && <div className="loading">Loading...</div>}
-                <MessageList messages={messages} />
+                <MessageList messages={messages} myId={myId} />
             </div>
             <div id='divider'></div>
             <MessageForm onMessageSubmit={handleMessageSubmit} myId={myId} accessToken={accessToken} imageFile={imageFile} setImageFile={setImageFile} />
